+++ src/main/java/services/UserService.java
@@ -10,7 +10,7 @@ public class UserService {
     public UserService() {
     }
 
-    public User findUser(int id) {
+    public User findUser(long id) {
         return usersDao.findById(id);
     }
